requirements.txt
Django==3.0.6
djangorestframework==3.11.0
django-rest-swagger==2.2.0
requests==2.10.0
psycopg2==2.8.5
djangorestframework-jwt==1.11.0